Project4: Got pieces to not be placed over each other with proper scores!
Created a brickus_placed attribute on each grabbed piece that keeps track
of whether it was snapped to the grid or not. It will only snap to the
grid if it passes a bunch of checks, so I know snapped pieces have
points that need to be removed.

I uncommented my code that prevented pieces from overlapping, and now
it works beautifully with the brickus_placed attribute check!