Project4: Got dragging and grid snapping to work with scrolling!
For piece movement, I needed to add window.page*Offset to the coordinates
for an absolute position relative to the top left corner of the document
(instead of just to the viewport - the visible upper-left corner).

Notably, I DIDN'T add the offsets to the grid-snapping logic, since
the underlying functions rely on visible gameboard pieces.