Brickus: Got a game to be created in db and users added to it
Created a Games model.
Changed the Users::query_user() to return a single record.
A game will be created if no games exist that are currently in progress.
If a user tries to join a game, it will add them to the game if there
is an opening, else it will say the game is full.